Moved AffineMatrixVisitor to its own files. Moved duplicated matrix calculations into Geometry::getTransform
