+++ src/axom/klee/Geometry.cpp
@@ -228,5 +228,47 @@ const std::string& Geometry::getBlueprintTopology() const
   return m_topology;
 }
 
+numerics::Matrix<double> Geometry::getTransform() const
+{
+  const auto identity4x4 = numerics::Matrix<double>::identity(4);
+  numerics::Matrix<double> transformation(identity4x4);
+  if(m_operator)
+  {
+    auto composite = std::dynamic_pointer_cast<const CompositeOperator>(m_operator);
+    if(composite)
+    {
+      // Concatenate the transformations
+
+      // Why don't we multiply the matrices in CompositeOperator::addOperator()?
+      // Why keep the matrices factored and multiply them here repeatedly?
+      // Combining them would also avoid this if-else logic.  BTNG
+      for(auto op : composite->getOperators())
+      {
+        // Use visitor pattern to extract the affine matrix from supported operators
+        AffineMatrixVisitor visitor;
+        op->accept(visitor);
+        if(!visitor.isValid())
+        {
+          continue;
+        }
+        const auto& matrix = visitor.getMatrix();
+        numerics::Matrix<double> res(identity4x4);
+        numerics::matrix_multiply(matrix, transformation, res);
+        transformation = res;
+      }
+    }
+    else
+    {
+      AffineMatrixVisitor visitor;
+      geometryOperator->accept(visitor);
+      if(visitor.isValid())
+      {
+        transformation = visitor.getMatrix();
+      }
+    }
+  }
+  return transformation;
+}
+
 }  // namespace klee
 }  // namespace axom
